requirements.txt
python-dotenv
pyyaml
anthropic
bs4
requests
mcp